# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: 匿名树洞社区 API
  version: 1.0.0
  description: 匿名发帖/评论、话题标签、点赞热度、每日精选、敏感词审核
servers:
  - url: http://localhost:19401
paths:
...
  /api/v1/posts:
    get:
      summary: 最新/标签/精选帖子列表
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
        - in: query
          name: tag_id
          schema: { type: integer }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        '200':
          description: 分页帖子列表
    post:
      summary: 发布帖子
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                title: { type: string }
                content: { type: string }
                images: { type: array, items: { type: string } }
                tags: { type: array, items: { type: string } }
      responses:
        '200':
          description: 帖子与命中信息
...
  /api/v1/posts/{id}:
    get:
      summary: 帖子详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 帖子详情
  /api/v1/posts/{id}/comments:
    get:
      summary: 帖子评论列表
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        '200':
          description: 分页评论列表
  /api/v1/comments:
    post:
      summary: 发表评论
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [postId, content]
              properties:
                postId: { type: integer }
                content: { type: string }
      responses:
        '200':
          description: 评论与命中信息
...